+++ gps/gps.go
@@ -2,67 +2,61 @@ package gps
 
 import (
 	"fmt"
-	"reflect"
-	"sync"
 
-	"github.com/kilianp07/AthleteIQBox/gps/nmea"
+	"github.com/kilianp07/AthleteIQBox/gps/reader"
+	"github.com/kilianp07/AthleteIQBox/gps/recorder"
 	"github.com/kilianp07/AthleteIQBox/utils"
 )
 
 var (
-	readers   = make(map[string]func() Reader)
-	readersMu sync.RWMutex
+	driver reader.Reader
+	saver  recorder.Recorder
 )
 
-func init() {
-	// Register the "nmea" reader
-	readersMu.Lock()
-	readers["nmea"] = func() Reader {
-		return nmea.New()
-	}
-	readersMu.Unlock()
-}
+func Configure(configuration Configuration) error {
 
-type Reader interface {
-	Configure() error
-	Start() error
-	Stop() error
-	RuntimeErr() chan error
-	Conf() any
-}
-
-type Configuration struct {
-	ID   string `json:"id"`
-	Conf any    `json:"conf"`
-}
-
-func New(conf Configuration) (Reader, error) {
-	readersMu.RLock()
-	constructor, ok := readers[conf.ID]
-	readersMu.RUnlock()
-	if !ok {
-		return nil, fmt.Errorf("invalid reader: %s", conf.ID)
-	}
-
-	r := constructor()
-
-	// Check if r.Conf returns a pointer
-	if reflect.ValueOf(r.Conf()).Kind() != reflect.Ptr {
-		return nil, fmt.Errorf("Conf() method of reader %s did not return a pointer", conf.ID)
-	}
-
-	decoder, err := utils.NewDecoder(r.Conf())
+	var conf Configuration
+	decoder, err := utils.NewDecoder(&conf)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create decoder: %w", err)
+		return fmt.Errorf("error creating decoder: %v", err)
 	}
 
-	if err := decoder.Decode(conf.Conf); err != nil {
-		return nil, fmt.Errorf("failed to decode configuration: %w", err)
+	err = decoder.Decode(configuration)
+	if err != nil {
+		return fmt.Errorf("error decoding configuration: %v", err)
 	}
 
-	if err := r.Configure(); err != nil {
-		return nil, err
+	if driver, err = reader.New(conf.Reader); err != nil {
+		return fmt.Errorf("error creating reader: %v", err)
 	}
 
-	return r, nil
+	if saver, err = recorder.New(conf.Recorder, driver.Position()); err != nil {
+		return fmt.Errorf("error creating recorder: %v", err)
+	}
+
+	return err
+}
+
+func Start() error {
+	if err := driver.Start(); err != nil {
+		return fmt.Errorf("error starting gps reader: %v", err)
+	}
+
+	if err := saver.Start(); err != nil {
+		return fmt.Errorf("error starting gps recorder: %v", err)
+	}
+
+	return nil
+}
+
+func Stop() error {
+	if err := driver.Stop(); err != nil {
+		return fmt.Errorf("error stopping gps reader: %v", err)
+	}
+
+	if err := saver.Stop(); err != nil {
+		return fmt.Errorf("error stopping gps recorder: %v", err)
+	}
+
+	return nil
 }