+++ gps/recorder/configuration.go
@@ -1,0 +1,6 @@
+package recorder
+
+type Configuration struct {
+	ID   string `json:"id"`
+	Conf any    `json:"conf"`
+}